% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loesungen.R
\name{Lösungen}
\alias{Lösungen}
\alias{Loesungen_1}
\alias{Lösungen_1}
\title{R-Skripte zur Lösung der Aufgaben im PsyMSc 1 Seminar, Sommersemester 2020

Mit diesen Funktionen können Sie die Lösungen zu den Übungen in den einzelnen Sitzungen aufrufen. Die Passwörter werden nach dem Abgabetermin im OLAT bekannt gegeben.}
\usage{
Loesungen_1(password)
}
\description{
R-Skripte zur Lösung der Aufgaben im PsyMSc 1 Seminar, Sommersemester 2020

Mit diesen Funktionen können Sie die Lösungen zu den Übungen in den einzelnen Sitzungen aufrufen. Die Passwörter werden nach dem Abgabetermin im OLAT bekannt gegeben.
}
