% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitzungen.R
\name{Sitzung_1}
\alias{Sitzung_1}
\title{Übungen im PsyMSc 1 Seminar, Sommersemester 2020}
\usage{
Sitzung_1()
}
\description{
Mit diesen Funktionen können Sie die Übungen zu den einzelnen Sitzungen aufrufen.
}
