% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitzungen.R
\name{Sitzungen}
\alias{Sitzungen}
\alias{Sitzung_1}
\alias{Sitzung_2}
\alias{Sitzung_3}
\alias{Sitzung_4}
\title{Übungen im PsyMSc 1 Seminar, Sommersemester 2020}
\usage{
Sitzung_1()

Sitzung_2()

Sitzung_3()

Sitzung_4()
}
\description{
Mit diesen Funktionen können Sie die Übungen zu den einzelnen Sitzungen aufrufen. Die Funktionen nehmen keine Argumente entgegen.
}
\details{
Durch die verschiedenen Funktionen können unterschiedliche Sitzungen aufgerufen werden. Dabei ist das Suffix _* immer die Nummer der jeweiligen Sitzung (z.B. Sitzung_1 für die 1. Sitzung). Die Inhalte wurden in \code{learnr} umgesetzt und das Kompilieren der Dateien findet anhand dieses Pakets statt.
}
\seealso{
Loesungen
}
