% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StressAtWork}
\alias{StressAtWork}
\title{Stress At Work Datensatz}
\format{
An object of class \code{data.frame} with 305 rows and 34 columns.
}
\source{
Der Datensatz `StressAtWork` ist eine Zusammenstellung aus mehreren Studien der Arbeits- und Organisationspsychologie Abteilung der Goethe-Universität, in welchen Call-Center-Mitarbeiter untersucht wurden.
}
\usage{
StressAtWork
}
\description{
Selbstberichte von Call-Center Mitarbeitern zu Geschlecht, Zeitdruck, arbeitsorganisationalen Problemen, psychosomatischen Beschwerden, emotionaler Erschöpfung und Leistungserfüllung, welche im Rahmen der Stressforschungs im Arbeitskontext von Relevanz sind.
}
\details{
@format Geschlecht der Probanden (`sex`, `1`=weiblich, `2`=männlich) sowie ausgewählte Messungen der Variablen _Zeitdruck_ (`zd1`, `zd2` und `zd6`) und _Arbeitsorganisationale Probleme_ (`aop3`, `aop4` und `aop8`) aus dem Instrument zu stressbezogenen Tätigkeitsanalyse (ISTA) von Semmer, Zapf und Dunckel (1999), _Psychosomatische Beschwerden_ (auch Befindlichkeit: `bf1`,...,`bf20`) aus der Psychosomatische Bschwerdenliste von Mohr (1986) sowie Messungen zu Subskalen von Burnout: _Emotionale Erschöpfung_ (`bo1`, `bo6`, `b12` und `b19`) und _Leistungserfüllung_ (`bo7`, `bo8` und `bo21`) aus Maslachs Burnout-Inventar (Maslach & Jackson, 1986) in der deutschen Übersetzung von Büssing und Perrar (1992).
}
\keyword{datasets}
